feat(standalone-tests-reorg): relocate existing tests into Enums/Services/Integration layout + Testbench/Pest setup and unit CI
