+++ phpunit.xml
@@ -8,8 +8,12 @@
 >
     <coverage/>
     <testsuites>
-        <testsuite name="optima">
+        <testsuite name="unit">
             <directory>./tests</directory>
+            <exclude>./tests/Integration</exclude>
+        </testsuite>
+        <testsuite name="integration">
+            <directory>./tests/Integration</directory>
         </testsuite>
     </testsuites>
     <source>
