+++ tests/Services/ParserTest.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Rudashi\Optima\Tests\Unit;
+namespace Rudashi\Optima\Tests\Services\ParserTest;
 
 use Rudashi\Optima\Exceptions\IncorrectValueException;
 use Rudashi\Optima\Services\Entity\Parser;
-use Rudashi\Optima\Tests\HelperClasses\FakeEnum;
+use Rudashi\Optima\Tests\Fixtures\FakeEnum;
 use Rudashi\Optima\Tests\TestCase;
 
 uses(TestCase::class);
